requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.116.1
greenlet==3.2.3
h11==0.16.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
pytest-mock==3.14.0
python-dotenv==1.1.1
python-multipart==0.0.9
PyYAML==6.0.2
requests==2.32.4
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.47.2
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
